+++ configureWebpack.js
@@ -97,10 +97,9 @@ try {reactPkg = require('react/package')} catch (e) {reactPkg = {version: 'NOT_F
  *      here to the Babel loader preset-env preset config.
  * @param {Object} [env.terserOptions] - options to spread onto / override defaults passed here to
  *      the Terser minification plugin for production builds.
- * @param {boolean} [env.loadAllBlueprintJsIcons] - false (default) to load only the several BlueprintJs
- *      icons the Hoist-React framework uses, and not load the ~500 icons that come with the BluePrint framework.
- *      Unlike the FontAwesome Icons that Hoist-React offers, the BluePrintJS icons, if all loaded, cannot be removed from the final build 
- *      via tree-shaking.  Sticking to the default "false" significantly reduces the size of the production build.
+ * @param {boolean} [env.loadAllBlueprintJsIcons] - false (default) to only load the BlueprintJs
+ *      icons required by Hoist React components, resulting in a much smaller bundle size. Set to
+ *      true if your app wishes to access all the BP icons (but consider FontAwesome instead!).
  */
 function configureWebpack(env) {
     if (!env.appCode) throw 'Missing required "appCode" config - cannot proceed';
@@ -137,8 +136,7 @@ function configureWebpack(env) {
         ],
         babelPresetEnvOptions = env.babelPresetEnvOptions || {},
         terserOptions = env.terserOptions || {},
-        buildDate = new Date(),
-        loadAllBlueprintJsIcons = env.loadAllBlueprintJsIcons === true;
+        buildDate = new Date();
 
     process.env.BABEL_ENV = prodBuild ? 'production' : 'development';
     process.env.NODE_ENV = prodBuild ? 'production' : 'development';
@@ -174,6 +172,10 @@ function configureWebpack(env) {
     const hoistPath = inlineHoist ?
         path.resolve(basePath, '../../hoist-react') :
         path.resolve(basePath, 'node_modules/@xh/hoist');
+
+    // Resolve Hoist-based path to replacement Blueprint icons (if available, requires HR >= v35.2.
+    const bpIconStubs = path.resolve(hoistPath, 'static/requiredBlueprintIcons.js'),
+        loadAllBlueprintJsIcons = env.loadAllBlueprintJsIcons === true || !bpIconStubs;
 
     // Tell webpack where to look for modules when resolving imports - this is the key to getting
     // inlineHoist mode to look in within the checked-out hoist-react project at hoistPath.
@@ -425,10 +427,10 @@ function configureWebpack(env) {
             // Clean (remove) the output directory before each run.
             new CleanWebpackPlugin(),
 
-            // Load only the several BlueprintJS icons that the Hoist-React framework uses.
+            // Load only the BlueprintJS icons used by Hoist-React components.
             !loadAllBlueprintJsIcons ? new webpack.NormalModuleReplacementPlugin(
               /.*\/generated\/iconSvgPaths.*/,
-              path.resolve(hoistPath, 'static/blueprintJsIconSvgPaths.js'),
+              bpIconStubs,
             ) : undefined,
 
             // Inject global constants at compile time.